requirements.txt
joblib==1.4.2
mlflow==2.22.0
mlflow_skinny==2.22.0
numpy==2.2.5
pandas==2.2.3
PyYAML==6.0.2
scikit_learn==1.6.1
typing_extensions==4.13.2
